Creacion completa del CRUD basico

+++ src/main/java/app/VilaExplorer/repository/FiestaTradicionRepository.java
@@ -5,8 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FiestaTradicionRepository extends JpaRepository <FiestaTradicion, Integer> {
-    // métodos personalizados, por ejemplo:
-    // List<FiestaTradicion> findByNombreContaining(String nombre);
-    // List<FiestaTradicion> findByAutorOrderByNombreAsc(Usuario autor);
+public interface FiestaTradicionRepository extends JpaRepository<FiestaTradicion, Long> {
 }
